requirements.txt
# LangChain & AI
langchain
langgraph
langchain-community
langchain-core
langchain-groq
langchain-anthropic
pydantic

# Data & Finance
yfinance
pandas
numpy

# Visualization
streamlit
plotly
matplotlib

# Machine Learning

scikit-learn

# Utilities
python-dotenv
IPython
vconsoleprint

-e .